Adding an option to reverse the Pico pinout.
The only downside (pun intended) to switching to the Pico as your go-to microcontroller, is that its pin numbers are printed on the underside. Other boards are more practical in that sense and easier to use as an educational tool, as those boards more clearly indicate what all pins are for. This inconvenience is remedied by soldering the pins so that the board sits on the breadboard with the underside facing up. This change adds the option to reverse the pinout to reflect that configuration.